# go/openapi.yaml
openapi: 3.0.1
info:
  title: Choreo BYOC Examples
  description: This repository contains demo API implemented in different frameworks to showcase capabilities of choreo.
  version: '0.1'
paths:
  /proxy/:
    post:
      description: Auto generated using Swagger Inspector
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                args:
                  type: string
                host:
                  type: string
            examples:
              '0':
                value: |-
                  {
                      "host": "http://postman-echo.com",
                      "args": "get?foo1=bar1&foo2=bar2"
                  }
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            text/html; charset=utf-8:
              schema:
                type: string
              examples: {}
  /healthz/:
    get:
      description: Auto generated using Swagger Inspector
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            application/json:
              schema:
                type: object
                properties:
                  healthy:
                    type: boolean
              examples:
                '0':
                  value: |
                    {"healthy":true}
  /:
    get:
      description: Auto generated using Swagger Inspector
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            application/json:
              schema:
                type: object
                properties:
                  healthy:
                    type: boolean
              examples:
                '0':
                  value: |
                    {"active":true}
  /hello/:
    get:
      description: Auto generated using Swagger Inspector
      parameters:
        - name: name
          in: query
          schema:
            type: string
          example: world
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            text/html; charset=utf-8:
              schema:
                type: string
              examples: {}
  /crash/:
    get:
      description: Terminates the server immediately to simulate a container crash (for testing purposes)
      responses:
        '200':
          description: Response sent before server terminates
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Server crashing now...
              examples:
                '0':
                  value: |
                    {"message": "Server crashing now..."}
  /status/:
    get:
      description: Returns server status, memory metrics, and runtime information
      responses:
        '200':
          description: Server status information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  server:
                    type: object
                    properties:
                      status:
                        type: string
                        example: running
                      uptime:
                        type: string
                        example: 1h23m45s
                      startTime:
                        type: string
                        format: date-time
                        example: '2025-11-11T10:30:00Z'
                  memory:
                    type: object
                    properties:
                      allocatedMB:
                        type: integer
                        example: 12
                      totalAllocMB:
                        type: integer
                        example: 25
                      sysMB:
                        type: integer
                        example: 50
                      numGC:
                        type: integer
                        example: 5
                  runtime:
                    type: object
                    properties:
                      numGoroutines:
                        type: integer
                        example: 10
                      goVersion:
                        type: string
                        example: go1.21.0
                      numCPU:
                        type: integer
                        example: 8
                  endpoints:
                    type: array
                    items:
                      type: string
                    example: ["/", "/healthz/", "/hello/", "/status/", "/proxy/"]
              examples:
                '0':
                  value: |
                    {
                      "server": {
                        "status": "running",
                        "uptime": "2h15m30s",
                        "startTime": "2025-11-11T10:30:00Z"
                      },
                      "memory": {
                        "allocatedMB": 12,
                        "totalAllocMB": 25,
                        "sysMB": 50,
                        "numGC": 5
                      },
                      "runtime": {
                        "numGoroutines": 10,
                        "goVersion": "go1.21.0",
                        "numCPU": 8
                      },
                      "endpoints": ["/", "/healthz/", "/hello/", "/status/", "/proxy/"]
                    }
        '405':
          description: Method not allowed - only GET is supported
          content:
            text/plain:
              schema:
                type: string
              examples:
                '0':
                  value: Method not allowed
  /config/:
    get:
      description: Returns the current application configuration including server settings, proxy defaults, and feature flags
      responses:
        '200':
          description: Current configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  server:
                    type: object
                    properties:
                      port:
                        type: integer
                        example: 9090
                      logLevel:
                        type: string
                        example: info
                  proxy:
                    type: object
                    properties:
                      defaultHost:
                        type: string
                        example: http://postman-echo.com
                      defaultPath:
                        type: string
                        example: get?foo1=bar1&foo2=bar2
                      requestTimeoutSeconds:
                        type: integer
                        example: 30
                  features:
                    type: object
                    properties:
                      enableStatusEndpoint:
                        type: boolean
                        example: true
                      enableConfigEndpoint:
                        type: boolean
                        example: true
                  configSource:
                    type: object
                    properties:
                      filePathEnvVar:
                        type: string
                        example: /config/app-config.json
                      loadedFromFile:
                        type: boolean
                        example: true
              examples:
                '0':
                  value: |
                    {
                      "server": {
                        "port": 9090,
                        "logLevel": "info"
                      },
                      "proxy": {
                        "defaultHost": "http://postman-echo.com",
                        "defaultPath": "get?foo1=bar1&foo2=bar2",
                        "requestTimeoutSeconds": 30
                      },
                      "features": {
                        "enableStatusEndpoint": true,
                        "enableConfigEndpoint": true
                      },
                      "configSource": {
                        "filePathEnvVar": "/config/app-config.json",
                        "loadedFromFile": true
                      }
                    }
        '405':
          description: Method not allowed - only GET is supported
          content:
            text/plain:
              schema:
                type: string
              examples:
                '0':
                  value: Method not allowed